created a seperated component of snackbar
