+++ client/src/interfaces/snackbar.props.interface.tsx
@@ -1,0 +1,5 @@
+export interface snackBarProps {
+    openSnackBar: boolean;
+    handleClose: () => void
+
+}
